Fast API & Chatbot/requirement.txt
fastapi
pydantic
uvicorn
paho-mqtt
langchain
langchain-google-genai
supabase
requests
typing
re
